refactor: separate Navbar
