+++ src/pages/_app.tsx
@@ -8,7 +8,7 @@ import { Noto_Sans_TC } from "next/font/google";
 
 import { Provider } from "react-redux";
 import { store } from "@/common/redux/store";
-import Navbar from "@/components/Navbar";
+import Navbar from "@/components/navbar";
 
 export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
   getLayout?: (page: ReactElement) => ReactNode;
